+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. This projec
 
 ## 1.0.13 (in progress)
 
++ [#13](https://github.com/luyadev/luya-testsuite/issues/13) Convert integer and boolean rule types into integer and boolean column types.
+
 ## 1.0.12 (8. October 2018)
 
 + [#12](https://github.com/luyadev/luya-testsuite/pull/12) Fixed issue with NgRestTestCase - Added missing is_api_user to User schema fixture